template - refactor (index.ts)
