--- code/compiler/template.esm/src/common.mts
@@ -1,1 +1,0 @@
-export * from './common/index.mjs';
