+++ code/compiler/template.esm/src/test/index.ts
@@ -1,3 +1,3 @@
 export { expect, expectError } from 'sys.test';
 export { describe, it } from 'vitest';
-export * from '../common.mjs';
+export * from '../common';
